refactor(action) `@alwatr/on` package has been removed and replaced by `@alwatr/action`. `DirectiveBase` in `@alwatr/directive` has been renamed to `Directive`.
Migration guide:

  @alwatr/on → @alwatr/action
  alwatr-on="…" attribute → on-action="…"
  alwatrOn(id, handler) → onAction(id, handler)
  alwatrDispatch(id, payload) → dispatchAction(id, payload)
  registerAlwatrOnDirective() → registerActionDirective()
  EventSignalPayload → ActionSignalPayload
  AlwatrActionDirective → ActionDirective

  @alwatr/directive
  DirectiveBase → Directive
