+++ pkg/nanolib/action/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@alwatr/on",
+  "name": "@alwatr/action",
   "version": "9.11.1",
   "description": "Declarative DOM action-dispatch — bridge HTML attributes to typed signal handlers.",
   "license": "MPL-2.0",
@@ -8,9 +8,9 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/Alwatr/alwatr",
-    "directory": "pkg/nanolib/on"
+    "directory": "pkg/nanolib/action"
   },
-  "homepage": "https://github.com/Alwatr/alwatr/tree/next/pkg/nanolib/on#readme",
+  "homepage": "https://github.com/Alwatr/alwatr/tree/next/pkg/nanolib/action#readme",
   "bugs": "https://github.com/Alwatr/alwatr/issues",
   "exports": {
     ".": {
@@ -77,6 +77,7 @@
     "spa",
     "typescript",
     "ui-library",
+    "unidirectional-data-flow",
     "vanilla-js",
     "web-development"
   ]
