--- pkg/nanolib/on/src/main.ts
@@ -1,2 +1,0 @@
-export * from './method.js';
-export * from './directive.js';
